fixed 64 bit precision warning

+++ LBYouTubeView/LBYouTubeExtractor.m
@@ -104,7 +104,7 @@ NSInteger const LBYouTubePlayerExtractorErrorCodeNoJSONData   =    3;
             checkingResult = [videos objectAtIndex:0];
         }
         else if (self.quality == LBYouTubeVideoQualityMedium) {
-            unsigned int index = MIN(videos.count-1, 1U);
+            unsigned long index = MIN(videos.count-1, 1U);
             checkingResult= [videos objectAtIndex:index];
         }
         else {
